+++ admin_command.go
@@ -71,65 +71,67 @@ const (
 	_QUERY_END        int = 50
 )
 
-type adminCommand struct {
+// AdminCommand allows managing user access to the server
+type AdminCommand struct {
 	dataBuffer []byte
 	dataOffset int
 }
 
-func newAdminCommand(buf []byte) *adminCommand {
+// NewAdminCommand returns an AdminCommand object.
+func NewAdminCommand(buf []byte) *AdminCommand {
 	if buf == nil {
 		buf = make([]byte, 10*1024)
 	}
-	return &adminCommand{
+	return &AdminCommand{
 		dataBuffer: buf,
 		dataOffset: 8,
 	}
 }
 
-func (acmd *adminCommand) createUser(cluster *Cluster, policy *AdminPolicy, user string, password []byte, roles []string) Error {
+func (acmd *AdminCommand) createUser(conn *Connection, policy *AdminPolicy, user string, password []byte, roles []string) Error {
 	acmd.writeHeader(_CREATE_USER, 3)
 	acmd.writeFieldStr(_USER, user)
 	acmd.writeFieldBytes(_PASSWORD, password)
 	acmd.writeRoles(roles)
-	return acmd.executeCommand(cluster, policy)
+	return acmd.executeCommand(conn, policy)
 }
 
-func (acmd *adminCommand) dropUser(cluster *Cluster, policy *AdminPolicy, user string) Error {
+func (acmd *AdminCommand) dropUser(conn *Connection, policy *AdminPolicy, user string) Error {
 	acmd.writeHeader(_DROP_USER, 1)
 	acmd.writeFieldStr(_USER, user)
-	return acmd.executeCommand(cluster, policy)
+	return acmd.executeCommand(conn, policy)
 }
 
-func (acmd *adminCommand) setPassword(cluster *Cluster, policy *AdminPolicy, user string, password []byte) Error {
+func (acmd *AdminCommand) setPassword(conn *Connection, policy *AdminPolicy, user string, password []byte) Error {
 	acmd.writeHeader(_SET_PASSWORD, 2)
 	acmd.writeFieldStr(_USER, user)
 	acmd.writeFieldBytes(_PASSWORD, password)
-	return acmd.executeCommand(cluster, policy)
+	return acmd.executeCommand(conn, policy)
 }
 
-func (acmd *adminCommand) changePassword(cluster *Cluster, policy *AdminPolicy, user string, password []byte) Error {
+func (acmd *AdminCommand) changePassword(conn *Connection, policy *AdminPolicy, user string, oldPass, newPass []byte) Error {
 	acmd.writeHeader(_CHANGE_PASSWORD, 3)
 	acmd.writeFieldStr(_USER, user)
-	acmd.writeFieldBytes(_OLD_PASSWORD, cluster.Password())
-	acmd.writeFieldBytes(_PASSWORD, password)
-	return acmd.executeCommand(cluster, policy)
+	acmd.writeFieldBytes(_OLD_PASSWORD, oldPass)
+	acmd.writeFieldBytes(_PASSWORD, newPass)
+	return acmd.executeCommand(conn, policy)
 }
 
-func (acmd *adminCommand) grantRoles(cluster *Cluster, policy *AdminPolicy, user string, roles []string) Error {
+func (acmd *AdminCommand) grantRoles(conn *Connection, policy *AdminPolicy, user string, roles []string) Error {
 	acmd.writeHeader(_GRANT_ROLES, 2)
 	acmd.writeFieldStr(_USER, user)
 	acmd.writeRoles(roles)
-	return acmd.executeCommand(cluster, policy)
+	return acmd.executeCommand(conn, policy)
 }
 
-func (acmd *adminCommand) revokeRoles(cluster *Cluster, policy *AdminPolicy, user string, roles []string) Error {
+func (acmd *AdminCommand) revokeRoles(conn *Connection, policy *AdminPolicy, user string, roles []string) Error {
 	acmd.writeHeader(_REVOKE_ROLES, 2)
 	acmd.writeFieldStr(_USER, user)
 	acmd.writeRoles(roles)
-	return acmd.executeCommand(cluster, policy)
+	return acmd.executeCommand(conn, policy)
 }
 
-func (acmd *adminCommand) createRole(cluster *Cluster, policy *AdminPolicy, roleName string, privileges []Privilege, whitelist []string, readQuota, writeQuota uint32) Error {
+func (acmd *AdminCommand) createRole(conn *Connection, policy *AdminPolicy, roleName string, privileges []Privilege, whitelist []string, readQuota, writeQuota uint32) Error {
 	fieldCount := 1
 	if len(privileges) > 1 {
 		fieldCount++
@@ -168,34 +170,34 @@ func (acmd *adminCommand) createRole(cluster *Cluster, policy *AdminPolicy, role
 		acmd.writeFieldUint32(_WRITE_QUOTA, writeQuota)
 	}
 
-	return acmd.executeCommand(cluster, policy)
+	return acmd.executeCommand(conn, policy)
 }
 
-func (acmd *adminCommand) dropRole(cluster *Cluster, policy *AdminPolicy, roleName string) Error {
+func (acmd *AdminCommand) dropRole(conn *Connection, policy *AdminPolicy, roleName string) Error {
 	acmd.writeHeader(_DROP_ROLE, 1)
 	acmd.writeFieldStr(_ROLE, roleName)
-	return acmd.executeCommand(cluster, policy)
+	return acmd.executeCommand(conn, policy)
 }
 
-func (acmd *adminCommand) grantPrivileges(cluster *Cluster, policy *AdminPolicy, roleName string, privileges []Privilege) Error {
+func (acmd *AdminCommand) grantPrivileges(conn *Connection, policy *AdminPolicy, roleName string, privileges []Privilege) Error {
 	acmd.writeHeader(_GRANT_PRIVILEGES, 2)
 	acmd.writeFieldStr(_ROLE, roleName)
 	if err := acmd.writePrivileges(privileges); err != nil {
 		return err
 	}
-	return acmd.executeCommand(cluster, policy)
+	return acmd.executeCommand(conn, policy)
 }
 
-func (acmd *adminCommand) revokePrivileges(cluster *Cluster, policy *AdminPolicy, roleName string, privileges []Privilege) Error {
+func (acmd *AdminCommand) revokePrivileges(conn *Connection, policy *AdminPolicy, roleName string, privileges []Privilege) Error {
 	acmd.writeHeader(_REVOKE_PRIVILEGES, 2)
 	acmd.writeFieldStr(_ROLE, roleName)
 	if err := acmd.writePrivileges(privileges); err != nil {
 		return err
 	}
-	return acmd.executeCommand(cluster, policy)
+	return acmd.executeCommand(conn, policy)
 }
 
-func (acmd *adminCommand) setWhitelist(cluster *Cluster, policy *AdminPolicy, roleName string, whitelist []string) Error {
+func (acmd *AdminCommand) setWhitelist(conn *Connection, policy *AdminPolicy, roleName string, whitelist []string) Error {
 	fieldCount := 1
 	if len(whitelist) > 0 {
 		fieldCount++
@@ -205,21 +207,22 @@ func (acmd *adminCommand) setWhitelist(cluster *Cluster, policy *AdminPolicy, ro
 	if len(whitelist) > 0 {
 		acmd.writeWhitelist(whitelist)
 	}
-	return acmd.executeCommand(cluster, policy)
+	return acmd.executeCommand(conn, policy)
 }
 
-func (acmd *adminCommand) setQuotas(cluster *Cluster, policy *AdminPolicy, roleName string, readQuota, writeQuota uint32) Error {
+func (acmd *AdminCommand) setQuotas(conn *Connection, policy *AdminPolicy, roleName string, readQuota, writeQuota uint32) Error {
 	acmd.writeHeader(_SET_QUOTAS, 3)
 	acmd.writeFieldStr(_ROLE, roleName)
 	acmd.writeFieldUint32(_READ_QUOTA, readQuota)
 	acmd.writeFieldUint32(_WRITE_QUOTA, writeQuota)
-	return acmd.executeCommand(cluster, policy)
+	return acmd.executeCommand(conn, policy)
 }
 
-func (acmd *adminCommand) queryUser(cluster *Cluster, policy *AdminPolicy, user string) (*UserRoles, Error) {
+// QueryUser returns user information.
+func (acmd *AdminCommand) QueryUser(conn *Connection, policy *AdminPolicy, user string) (*UserRoles, Error) {
 	acmd.writeHeader(_QUERY_USERS, 1)
 	acmd.writeFieldStr(_USER, user)
-	list, err := acmd.readUsers(cluster, policy)
+	list, err := acmd.readUsers(conn, policy)
 	if err != nil {
 		return nil, err
 	}
@@ -231,19 +234,21 @@ func (acmd *adminCommand) queryUser(cluster *Cluster, policy *AdminPolicy, user 
 	return nil, nil
 }
 
-func (acmd *adminCommand) queryUsers(cluster *Cluster, policy *AdminPolicy) ([]*UserRoles, Error) {
+// QueryUsers returns user information for all users.
+func (acmd *AdminCommand) QueryUsers(conn *Connection, policy *AdminPolicy) ([]*UserRoles, Error) {
 	acmd.writeHeader(_QUERY_USERS, 0)
-	list, err := acmd.readUsers(cluster, policy)
+	list, err := acmd.readUsers(conn, policy)
 	if err != nil {
 		return nil, err
 	}
 	return list, nil
 }
 
-func (acmd *adminCommand) queryRole(cluster *Cluster, policy *AdminPolicy, roleName string) (*Role, Error) {
+// QueryRole returns role information.
+func (acmd *AdminCommand) QueryRole(conn *Connection, policy *AdminPolicy, roleName string) (*Role, Error) {
 	acmd.writeHeader(_QUERY_ROLES, 1)
 	acmd.writeFieldStr(_ROLE, roleName)
-	list, err := acmd.readRoles(cluster, policy)
+	list, err := acmd.readRoles(conn, policy)
 	if err != nil {
 		return nil, err
 	}
@@ -255,16 +260,17 @@ func (acmd *adminCommand) queryRole(cluster *Cluster, policy *AdminPolicy, roleN
 	return nil, nil
 }
 
-func (acmd *adminCommand) queryRoles(cluster *Cluster, policy *AdminPolicy) ([]*Role, Error) {
+// QueryRoles returns role information for all roles.
+func (acmd *AdminCommand) QueryRoles(conn *Connection, policy *AdminPolicy) ([]*Role, Error) {
 	acmd.writeHeader(_QUERY_ROLES, 0)
-	list, err := acmd.readRoles(cluster, policy)
+	list, err := acmd.readRoles(conn, policy)
 	if err != nil {
 		return nil, err
 	}
 	return list, nil
 }
 
-func (acmd *adminCommand) writeRoles(roles []string) {
+func (acmd *AdminCommand) writeRoles(roles []string) {
 	offset := acmd.dataOffset + int(_FIELD_HEADER_SIZE)
 	acmd.dataBuffer[offset] = byte(len(roles))
 	offset++
@@ -280,7 +286,7 @@ func (acmd *adminCommand) writeRoles(roles []string) {
 	acmd.dataOffset = offset
 }
 
-func (acmd *adminCommand) writePrivileges(privileges []Privilege) Error {
+func (acmd *AdminCommand) writePrivileges(privileges []Privilege) Error {
 	offset := acmd.dataOffset + int(_FIELD_HEADER_SIZE)
 	acmd.dataBuffer[offset] = byte(len(privileges))
 	offset++
@@ -320,7 +326,7 @@ func (acmd *adminCommand) writePrivileges(privileges []Privilege) Error {
 	return nil
 }
 
-func (acmd *adminCommand) writeWhitelist(whitelist []string) {
+func (acmd *AdminCommand) writeWhitelist(whitelist []string) {
 	offset := acmd.dataOffset + int(_FIELD_HEADER_SIZE)
 
 	comma := false
@@ -340,13 +346,13 @@ func (acmd *adminCommand) writeWhitelist(whitelist []string) {
 	acmd.dataOffset = offset
 }
 
-func (acmd *adminCommand) writeSize() {
+func (acmd *AdminCommand) writeSize() {
 	// Write total size of message which is the current offset.
 	var size = int64(acmd.dataOffset-8) | (_MSG_VERSION << 56) | (_MSG_TYPE << 48)
 	binary.BigEndian.PutUint64(acmd.dataBuffer[0:], uint64(size))
 }
 
-func (acmd *adminCommand) writeHeader(command byte, fieldCount int) {
+func (acmd *AdminCommand) writeHeader(command byte, fieldCount int) {
 	// Authenticate header is almost all zeros
 	for i := acmd.dataOffset; i < acmd.dataOffset+16; i++ {
 		acmd.dataBuffer[i] = 0
@@ -356,25 +362,25 @@ func (acmd *adminCommand) writeHeader(command byte, fieldCount int) {
 	acmd.dataOffset += 16
 }
 
-func (acmd *adminCommand) writeFieldStr(id byte, str string) {
+func (acmd *AdminCommand) writeFieldStr(id byte, str string) {
 	len := copy(acmd.dataBuffer[acmd.dataOffset+int(_FIELD_HEADER_SIZE):], str)
 	acmd.writeFieldHeader(id, len)
 	acmd.dataOffset += len
 }
 
-func (acmd *adminCommand) writeFieldBytes(id byte, bytes []byte) {
+func (acmd *AdminCommand) writeFieldBytes(id byte, bytes []byte) {
 	copy(acmd.dataBuffer[acmd.dataOffset+int(_FIELD_HEADER_SIZE):], bytes)
 	acmd.writeFieldHeader(id, len(bytes))
 	acmd.dataOffset += len(bytes)
 }
 
-func (acmd *adminCommand) writeFieldUint32(id byte, size uint32) {
+func (acmd *AdminCommand) writeFieldUint32(id byte, size uint32) {
 	acmd.writeFieldHeader(id, 4)
 	binary.BigEndian.PutUint32(acmd.dataBuffer[acmd.dataOffset:], size)
 	acmd.dataOffset += 4
 }
 
-func (acmd *adminCommand) writeFieldHeader(id byte, size int) {
+func (acmd *AdminCommand) writeFieldHeader(id byte, size int) {
 	// Buffer.Int32ToBytes(int32(size+1), acmd.dataBuffer, acmd.dataOffset)
 	binary.BigEndian.PutUint32(acmd.dataBuffer[acmd.dataOffset:], uint32(size+1))
 
@@ -383,61 +389,48 @@ func (acmd *adminCommand) writeFieldHeader(id byte, size int) {
 	acmd.dataOffset++
 }
 
-func (acmd *adminCommand) executeCommand(cluster *Cluster, policy *AdminPolicy) Error {
+func (acmd *AdminCommand) executeCommand(conn *Connection, policy *AdminPolicy) Error {
 	acmd.writeSize()
-	node, err := cluster.GetRandomNode()
-	if err != nil {
-		return err
-	}
 	timeout := 1 * time.Second
 	if policy != nil && policy.Timeout > 0 {
 		timeout = policy.Timeout
 	}
 
-	node.tendConnLock.Lock()
-	defer node.tendConnLock.Unlock()
-
-	if err = node.initTendConn(timeout); err != nil {
+	if err := conn.SetTimeout(time.Now().Add(timeout), timeout); err != nil {
 		return err
 	}
 
-	conn := node.tendConn
-	if _, err = conn.Write(acmd.dataBuffer[:acmd.dataOffset]); err != nil {
+	if _, err := conn.Write(acmd.dataBuffer[:acmd.dataOffset]); err != nil {
 		return err
 	}
 
-	if _, err = conn.Read(acmd.dataBuffer, _HEADER_SIZE); err != nil {
+	if _, err := conn.Read(acmd.dataBuffer, _HEADER_SIZE); err != nil {
 		return err
 	}
 
 	result := acmd.dataBuffer[_RESULT_CODE]
 	if result != 0 {
-		return newCustomNodeError(node, types.ResultCode(result))
+		if conn.node != nil {
+			return newCustomNodeError(conn.node, types.ResultCode(result))
+		}
+		return newError(types.ResultCode(result))
 	}
 
 	return nil
 }
 
-func (acmd *adminCommand) readUsers(cluster *Cluster, policy *AdminPolicy) ([]*UserRoles, Error) {
+func (acmd *AdminCommand) readUsers(conn *Connection, policy *AdminPolicy) ([]*UserRoles, Error) {
 	acmd.writeSize()
-	node, err := cluster.GetRandomNode()
-	if err != nil {
-		return nil, err
-	}
 	timeout := 1 * time.Second
 	if policy != nil && policy.Timeout > 0 {
 		timeout = policy.Timeout
 	}
 
-	node.tendConnLock.Lock()
-	defer node.tendConnLock.Unlock()
-
-	if err = node.initTendConn(timeout); err != nil {
+	if err := conn.SetTimeout(time.Now().Add(timeout), timeout); err != nil {
 		return nil, err
 	}
 
-	conn := node.tendConn
-	if _, err = conn.Write(acmd.dataBuffer[:acmd.dataOffset]); err != nil {
+	if _, err := conn.Write(acmd.dataBuffer[:acmd.dataOffset]); err != nil {
 		return nil, err
 	}
 
@@ -447,12 +440,15 @@ func (acmd *adminCommand) readUsers(cluster *Cluster, policy *AdminPolicy) ([]*U
 	}
 
 	if status > 0 {
-		return nil, newCustomNodeError(node, types.ResultCode(status))
+		if conn.node != nil {
+			return nil, newCustomNodeError(conn.node, types.ResultCode(status))
+		}
+		return nil, newError(types.ResultCode(status))
 	}
 	return list, nil
 }
 
-func (acmd *adminCommand) readUserBlocks(conn *Connection) (status int, rlist []*UserRoles, err Error) {
+func (acmd *AdminCommand) readUserBlocks(conn *Connection) (status int, rlist []*UserRoles, err Error) {
 
 	var list []*UserRoles
 
@@ -483,7 +479,7 @@ func (acmd *adminCommand) readUserBlocks(conn *Connection) (status int, rlist []
 	return status, rlist, nil
 }
 
-func (acmd *adminCommand) parseUsers(receiveSize int) (int, []*UserRoles, Error) {
+func (acmd *AdminCommand) parseUsers(receiveSize int) (int, []*UserRoles, Error) {
 	acmd.dataOffset = 0
 	list := make([]*UserRoles, 0, 100)
 
@@ -539,7 +535,7 @@ func (acmd *adminCommand) parseUsers(receiveSize int) (int, []*UserRoles, Error)
 	return 0, list, nil
 }
 
-func (acmd *adminCommand) parseRoles(userRoles *UserRoles) {
+func (acmd *AdminCommand) parseRoles(userRoles *UserRoles) {
 	size := int(acmd.dataBuffer[acmd.dataOffset] & 0xFF)
 	acmd.dataOffset++
 	userRoles.Roles = make([]string, 0, size)
@@ -553,7 +549,7 @@ func (acmd *adminCommand) parseRoles(userRoles *UserRoles) {
 	}
 }
 
-func (acmd *adminCommand) parseInfo() []int {
+func (acmd *AdminCommand) parseInfo() []int {
 	size := int(acmd.dataBuffer[acmd.dataOffset] & 0xFF)
 	acmd.dataOffset++
 	list := make([]int, 0, size)
@@ -576,26 +572,18 @@ func hashPassword(password string) ([]byte, Error) {
 	return []byte(hashedPassword), nil
 }
 
-func (acmd *adminCommand) readRoles(cluster *Cluster, policy *AdminPolicy) ([]*Role, Error) {
+func (acmd *AdminCommand) readRoles(conn *Connection, policy *AdminPolicy) ([]*Role, Error) {
 	acmd.writeSize()
-	node, err := cluster.GetRandomNode()
-	if err != nil {
-		return nil, err
-	}
 	timeout := 1 * time.Second
 	if policy != nil && policy.Timeout > 0 {
 		timeout = policy.Timeout
 	}
 
-	node.tendConnLock.Lock()
-	defer node.tendConnLock.Unlock()
-
-	if err = node.initTendConn(timeout); err != nil {
+	if err := conn.SetTimeout(time.Now().Add(timeout), timeout); err != nil {
 		return nil, err
 	}
 
-	conn := node.tendConn
-	if _, err = conn.Write(acmd.dataBuffer[:acmd.dataOffset]); err != nil {
+	if _, err := conn.Write(acmd.dataBuffer[:acmd.dataOffset]); err != nil {
 		return nil, err
 	}
 
@@ -605,12 +593,15 @@ func (acmd *adminCommand) readRoles(cluster *Cluster, policy *AdminPolicy) ([]*R
 	}
 
 	if status > 0 {
-		return nil, newCustomNodeError(node, types.ResultCode(status))
+		if conn.node != nil {
+			return nil, newCustomNodeError(conn.node, types.ResultCode(status))
+		}
+		return nil, newError(types.ResultCode(status))
 	}
 	return list, nil
 }
 
-func (acmd *adminCommand) readRoleBlocks(conn *Connection) (status int, rlist []*Role, err Error) {
+func (acmd *AdminCommand) readRoleBlocks(conn *Connection) (status int, rlist []*Role, err Error) {
 
 	var list []*Role
 
@@ -641,7 +632,7 @@ func (acmd *adminCommand) readRoleBlocks(conn *Connection) (status int, rlist []
 	return status, rlist, nil
 }
 
-func (acmd *adminCommand) parseRolesFull(receiveSize int) (int, []*Role, Error) {
+func (acmd *AdminCommand) parseRolesFull(receiveSize int) (int, []*Role, Error) {
 	acmd.dataOffset = 0
 
 	var list []*Role
@@ -697,7 +688,7 @@ func (acmd *adminCommand) parseRolesFull(receiveSize int) (int, []*Role, Error) 
 	return 0, list, nil
 }
 
-func (acmd *adminCommand) parsePrivileges(role *Role) {
+func (acmd *AdminCommand) parsePrivileges(role *Role) {
 	size := int(acmd.dataBuffer[acmd.dataOffset] & 0xFF)
 	acmd.dataOffset++
 	role.Privileges = make([]Privilege, 0, size)
@@ -722,7 +713,7 @@ func (acmd *adminCommand) parsePrivileges(role *Role) {
 	}
 }
 
-func (acmd *adminCommand) parseWhitelist(length int) []string {
+func (acmd *AdminCommand) parseWhitelist(length int) []string {
 	list := []string{}
 	begin := acmd.dataOffset
 	max := begin + length